requirements.txt
streamlit>=1.29.0
google-generativeai>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
faker>=20.0.0
logging>=0.4.9.6
datetime
json5>=0.9.0
typing-extensions>=4.8.0
faker>=20.0.0
google-genai>=1.38.0